# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students/{id}": {
      "get": {
        "description": "Returns the full student record including all linked guardians. Out-of-scope or cross-school access returns **404** (not 403) to avoid leaking student existence. Guardians are loaded eagerly and sorted by creation time (primary first is not guaranteed — check `is_primary`). Does NOT write an audit log.",
        "operationId": "StudentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetailEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student not found or out of scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a student by ID (with guardians)",
        "tags": [
          "Students"
        ]
      },
      "patch": {
        "description": "Updates an in-scope student. Caller role determines allowed fields:\n- **principal / vice_principal**: all fields in `UpdateStudentDto` (bio + capacity + notes).\n- **teacher**: capacity fields + `notes` only (`UpdateStudentByTeacherDto`). Teacher must be the **primary teacher** on at least one of the student's halaqat; otherwise 403. Submitting bio fields as a teacher returns 400.\n\nCapacity values out of range (hifz 0–20, near 0–50, far 0–100) return 400. Audit: `student.update` with `oldValues`/`newValues` of changed fields only.",
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or capacity out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Teacher not primary on any of the student's halaqat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student not found or out of scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a student (BR-STU-02)",
        "tags": [
          "Students"
        ]
      },
      "delete": {
        "description": "Soft-deletes the student (sets `deleted_at`). The student no longer appears in list results unless `include_deleted=true` is passed. Guardian links are preserved and can be restored. Audit: `student.delete`.",
        "operationId": "StudentsController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Student deleted (no body)"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student not found or out of scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a student (BR-STU-03)",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/students/{id}/guardians": {
      "get": {
        "description": "Returns all guardians linked to the student, sorted by creation time (ascending). Caller must pass the `StudentScopeGuard` — out-of-scope or cross-school access returns 404. Does NOT write an audit log.",
        "operationId": "StudentGuardiansController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardianListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student not found or out of scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List guardians for a student",
        "tags": [
          "Student Guardians"
        ]
      },
      "post": {
        "description": "Links an existing user to the student as a guardian via one of two flows:\n\n**ID flow** (`guardian_user_id`): the user must exist in the same school. Returns 404 if not found.\n\n**Email flow** (`email`): looks up an existing user by email in the same school. Returns 404 if not found, 409 if the account exists but is deactivated.\n\nBoth flows call `ensureRoleBySlug(\"parent\", ...)` — idempotent, adds the role if missing.\n\n**`is_primary` rules:**\n1. First guardian on the student → forced to `is_primary=true`.\n2. Subsequent guardian with `is_primary=true` → unsets the existing primary first.\n3. Subsequent guardian with `is_primary=false` (or omitted) → leaves existing primary alone.\n\nAudit: `student.guardian.link`.",
        "operationId": "StudentGuardiansController_link",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkGuardianDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardianEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or guardian already linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student or guardian user not found in this school",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Guardian user account is deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Link a guardian to a student (BR-STU-05)",
        "tags": [
          "Student Guardians"
        ]
      }
    },
    "/students/{id}/guardians/{guardianUserId}": {
      "patch": {
        "description": "Updates `relation`, `is_primary`, or `can_pickup` for an existing guardian link.\n\n**`is_primary` rules:**\n- `is_primary=true`: unsets the current primary, then sets this guardian as primary. Audit: `student.guardian.update` + implicit `student.guardian.primary_promoted` for the old primary.\n- `is_primary=false`: always rejected with **400** — demoting a primary is not permitted directly. Instead, promote a different guardian.\n\nAudit: `student.guardian.update`.",
        "operationId": "StudentGuardiansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "guardianUserId",
            "required": true,
            "in": "path",
            "description": "Guardian user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuardianDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardianEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "is_primary=false rejected or validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student or guardian link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a guardian link (BR-STU-06)",
        "tags": [
          "Student Guardians"
        ]
      },
      "delete": {
        "description": "Removes the guardian link. Behaviour depends on whether the guardian is the primary:\n\n- **Non-primary**: unlinked without side effects.\n- **Primary, others remain**: earliest-created remaining guardian is auto-promoted to primary. Audit: `student.guardian.primary_promoted`.\n- **Last guardian**: link removed. Audit: `student.orphaned`. A notification is dispatched to all vice_principals in the school via `NotificationService` (currently a logger stub — real delivery TBD).\n\nAudit: `student.guardian.unlink`.",
        "operationId": "StudentGuardiansController_unlink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "guardianUserId",
            "required": true,
            "in": "path",
            "description": "Guardian user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Guardian unlinked (no body)"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student or guardian link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Unlink a guardian from a student (BR-STU-07)",
        "tags": [
          "Student Guardians"
        ]
      }
    },
...
    "/attendance/students/{id}": {
      "patch": {
        "description": "Changes the status of one attendance row, recording who changed it, why (`modification_reason`), and the value it had before the first correction (`original_status`). Out-of-scope or cross-school rows return 404.",
        "operationId": "StudentAttendanceController_correct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Attendance row ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceDto"
                }
              }
            }
          },
          "400": {
            "description": "Status unchanged."
          },
          "404": {
            "description": "Not found or out of scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Correct a single attendance row",
        "tags": [
          "Attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Halaqa Backend API",
    "description": "Authentication, user management, and school operations for the Halaqa system. All JSON responses are wrapped in a uniform envelope: `{ \"code\": <http>, \"data\": ... }` on success, `{ \"code\": <http>, \"message\": \"...\" }` on errors. Refresh tokens travel as HttpOnly cookies on `/auth/*`; access tokens use Bearer auth.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "details": {
            "description": "Validation error list (only on 400 with multiple field errors).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "StudentDetailEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/StudentDetailResponse"
          },
          "warnings": {
            "example": [
              "id_number.checksum_invalid"
            ],
            "description": "Non-fatal warnings. Present only when id_number validation produced warnings.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
      "LinkGuardianDto": {
        "type": "object",
        "properties": {
          "guardian_user_id": {
            "type": "number",
            "example": 42,
            "description": "ID of an existing user. Mutually exclusive with `email`. Must belong to the same school."
          },
          "email": {
            "type": "string",
            "example": "parent@example.com",
            "format": "email",
            "description": "Email of an existing user. Mutually exclusive with `guardian_user_id`. Returns 404 if not found."
          },
          "relation": {
            "type": "string",
            "enum": [
              "father",
              "mother",
              "grandfather",
              "grandmother",
              "uncle",
              "aunt",
              "sibling",
              "other"
            ],
            "example": "father"
          },
          "is_primary": {
            "type": "boolean",
            "example": true,
            "description": "Exactly one guardian per student may be primary. If omitted on the first guardian, forced to true. Setting true unsets the current primary. Setting false on the current primary is rejected (400)."
          },
          "can_pickup": {
            "type": "boolean",
            "example": true,
            "description": "Whether the guardian is authorised to pick up the student. Defaults to true."
          }
        },
        "required": [
          "relation"
        ]
      },
...
      "StudentEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/StudentResponse"
          },
          "warnings": {
            "example": [
              "id_number.checksum_invalid"
            ],
            "description": "Non-fatal warnings. Present only when id_number validation produced warnings.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "يوسف محمد",
            "minLength": 2,
            "maxLength": 100
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "2015-06-01",
            "nullable": true
          },
          "join_date": {
            "type": "string",
            "format": "date",
            "example": "2024-09-01"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "graduated"
            ],
            "example": "active",
            "description": "Use POST /students/:id/graduate to set graduated; this field accepts any status for admin overrides."
          },
          "daily_hifz_pages_capacity": {
            "type": "number",
            "example": 1,
            "minimum": 0,
            "maximum": 20
          },
          "daily_near_pages_capacity": {
            "type": "number",
            "example": 2,
            "minimum": 0,
            "maximum": 50
          },
          "daily_far_pages_capacity": {
            "type": "number",
            "example": 5,
            "minimum": 0,
            "maximum": 100
          },
          "notes": {
            "type": "string",
            "example": "Needs extra support on Juz 30."
          },
          "photo_url": {
            "type": "string",
            "example": "https://cdn.example.com/photos/1.jpg"
          },
          "id_number": {
            "type": "object",
            "nullable": true,
            "example": "300123456",
            "maxLength": 20,
            "description": "National ID number. Set to null to clear a previously-set value (requires force_id_number_change: true). Changing an already-set value also requires the flag. A bad checksum is stored with a warning in the response."
          },
          "force_id_number_change": {
            "type": "boolean",
            "example": true,
            "description": "Required when changing or clearing an already-set id_number. Transient — never persisted."
          }
        }
      },
...
      "GuardianListEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuardianResponse"
            }
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
      "GuardianEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/GuardianResponse"
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
      "UpdateGuardianDto": {
        "type": "object",
        "properties": {
          "relation": {
            "type": "string",
            "enum": [
              "father",
              "mother",
              "grandfather",
              "grandmother",
              "uncle",
              "aunt",
              "sibling",
              "other"
            ],
            "example": "mother"
          },
          "is_primary": {
            "type": "boolean",
            "example": true,
            "description": "Set to true to promote this guardian to primary (unsets the current primary). Setting false is rejected with 400 — use another guardian's PATCH with is_primary=true instead."
          },
          "can_pickup": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "AttendanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1001
          },
          "student_id": {
            "type": "number",
            "example": 42
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-07-07"
          },
          "status": {
            "type": "string",
            "enum": [
              "present",
              "absent",
              "excused",
              "late"
            ],
            "example": "present"
          },
          "excuse_note": {
            "type": "object",
            "nullable": true,
            "example": "مريض"
          },
          "recorded_by": {
            "type": "object",
            "nullable": true,
            "example": 7,
            "description": "NULL = auto-seeded by the system."
          },
          "modified_by": {
            "type": "object",
            "nullable": true,
            "example": 7
          },
          "modification_reason": {
            "type": "object",
            "nullable": true,
            "example": "Marked present by mistake."
          },
          "original_status": {
            "type": "object",
            "nullable": true,
            "example": "present"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "student_id",
          "date",
          "status",
          "excuse_note",
          "recorded_by",
          "modified_by",
          "modification_reason",
          "original_status",
          "created_at"
        ]
      },
...
      "CorrectAttendanceDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "present",
              "absent",
              "excused",
              "late"
            ],
            "example": "excused"
          },
          "excuse_note": {
            "type": "string",
            "nullable": true,
            "example": "عذر طبي"
          },
          "modification_reason": {
            "type": "string",
            "example": "Marked present by mistake in the morning."
          }
        },
        "required": [
          "status",
          "modification_reason"
        ]
      },
...
    }
  }
}